requirements.txt
asgiref==3.4.1
Django==3.2.6
django-crispy-forms==1.13.0
django-recaptcha==2.0.6
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
